ipam, virt: update IPAMClaim with the pods allocated IPs
Integrate the IPAMClaim reconciliation (updating the IPAMClaim with the
pod's allocated IPs) after the pod's annotation is allocated for both
interconnect / non-interconnect scenarios.

This feature is only available for secondary networks of `layer2` and
`localnet` topologies.

A follow-up commit will introduce the actual code / unit tests of the
IPAMClaim reconciliation.

Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>