ipam, virt: do not release IPs if persistent IP addrs are used
This commit prevents the release of IP addresses back to the pool upon
the pod delete events when persistent IP addresses are being used (i.e.
the network selection element features an `ipam-claim-reference`).

The IPs will only be returned to the pool once the corresponding claim
is deleted - until then, the reservation should hold.

Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>