ipam, virt: sync the IP allocator using IPAMClaims
This commit syncs the IP allocator using IPAMClaims (e.g. it now uses
both live pods *and* IPAMClaims) when the network cluster controller
comes up (e.g. after a restart).

This way, we ensure that existing allocations to stopped VMs are not
re-assigned to other workloads.

This commit introduces this sync directly via the event handler, thus
ensuring the IP addresses are sync'ed for both interconnect /
non-interconnet deployment types.

It must be highlighted that a network controller only syncs IPAMClaim
for the network it manages.

Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>