Handle IP fragments in SGW mode
When OVS is operating in normal fragmentation mode, IP fragments
ingressing the OVS pipeline won't have their L4 fields set and node port
flows that match on them won't be hit.

In the best of cases, this results in packets forwarded via host with
the potential performance impact. In the worst of cases, a mix of
fragmented and non-fragmented packets for a connection forwarded via
host and directly to OVN respectively might result in a broken
connection as OVN just thinks each are different connections due to the
different NAT done at via host vs via OVN.

This change mitigates the issue sending IP fragments to conntrack on a
dedicated zone. The kernel implemetation of conntrack will reassemble
packets as long the netfilter kernel module is loaded. The reassembled
packets will be reinjected in table 0.

This change has a number of caveats:
* won't work for userspace datapath as the userspace implementation of
  conntrack does not reassemble packets. A more elaborate change in the
  pipeline is possible to properly support the userspace datapath but
  went for the simple approach first
* it is possible for the reassembled packet to be larger that the pod
  network MTU and OVN will send a ICMP needs frag back to the client as
  a result. However if fragmentation was ocurring in the first place, it
  could very well be possible that PMTUD is not in effect and that ICMP
  won't help in any way. This is unlikely to be a new issue in any case
  as the GR is most likely sending packets through conntrack as well.

PMTUD might not be in effect either because the client:
* disables it using IP_PMTUDISC_DONT or setting net.ipv4.ip_no_pmtu_disc
* uses UDP, IP_PMTUDISC_WANT and sends packets larger than the PMTU

This last case is used in the added test case.

Signed-off-by: Jaime Caamaño Ruiz <jcaamano@redhat.com>