helm: Add GitHub Test Lane
This commit adds a new lane to test the GitHub
workflow, exercising the creation of an OVN
Kubernetes cluster via Helm. It introduces a
script called contrib/kind-helm.sh, which can
be used instead of contrib/kind.sh.

Even though the test lane has a placeholder
in the matrix for ovn-interconnect, this is not
being implemented in this commit.

Signed-off-by: Flavio Fernandes <ffernandes@nvidia.com>