Copy health check functions to its own package
Health check is used both by egress ip and egress services controllers
and could potentially be used for more stuff in the future. It makes
sense for the code to live in its own package.

It's a copy and not a move due to the way egress service mocks those
functions on their tests. This will be cleaned up on a later commit.

Signed-off-by: Jaime Caamaño Ruiz <jcaamano@redhat.com>