fixes ipv6 control plane testing
fixes "Services of type NodePort: should listen on each host addresses",
the issue is that for IPv6 addresses the external container is
outside the range of the of the network that the OCI uses to setup the
enviroment. We use fc00:f853:ccd:e798::/64 as the OCI network subnet and
where using fc00:f853:ccd:e794:: for the external container which is
outside. Changing the external container to fc00:f853:ccd:e793:1111::
fixes the issue.

Additionally the subnet mask was hardcoded to "/32" regardless of ip
family. That is corrected here as well.

Unblocks
should listen on each host addresses
Should be allowed to node local cluster-networked endpoints by nodeport services with externalTrafficPolicy=local

Test "e2e ingress to host-networked pods traffic validation: Should be
allowed to node local host-networked endpoints by nodeport services on
IPV6 clusters and LGW mode" fails because

Fixes: https://github.com/ovn-org/ovn-kubernetes/issues/4136
Fixes: https://github.com/ovn-org/ovn-kubernetes/issues/4135

Signed-off-by: Jacob Tanenbaum <jtanenba@redhat.com>