fix for ipv6 control plane test for nodePorts
currently the test "Services when a nodePort service targeting a pod
with hostNetwork:true is created when tests are run towards the agnhost
echo service [It] queries to the nodePort service shall work for UDP on
IPV6 Shared gateway" fails because the test is clearing the ipv4 cache
instead of the ipv6 cache

Correcting the operation of this test to clear the ipv6 cache when the
service ip is ipv6. This fixes the issue with this test.

Fixes: https://github.com/ovn-org/ovn-kubernetes/issues/4138
Signed-off-by: Jacob Tanenbaum <jtanenba@redhat.com>