EIP: Add QoS rule at prio 103 to mark reply packet
This commit introduces qos rules on every node's switch
to mark packets emerging from EIP pods if they are
reply packets. These are the packets that should not be
reRouted to other node in the cluster. So basically we
mark the SYNACK packets that are matched on by the LRP we
saw in the previous commit. SYN packets are not marked
as these are egress traffic from EIP pods.

sample:
from-lport   103 (ip4.src==$a4548040316634674295 && ct.trk && ct.rpl) mark=42

This let's these two features (ETP=local + EIP)
to be used at the same time.

Signed-off-by: Surya Seetharaman <suryaseetharaman.9@gmail.com>