Load balancer failed to be created
Load balancer failed to be created because it took long time to
initialize nodeTracker in case there were lots of nodes. As the result,
service controller handler go routine timed out after 20 seconds and
returned error.

This commit increases HandlerSyncTimeout to 3 minutes, and it would be
a fatal error if service controller fails to start.

Signed-off-by: Yun Zhou <yunz@nvidia.com>