Introduce UNAVAILABLE health state
Expressing a situation where the node is alive with an apparently
functioning OVN dataplane (given that the health check itself is done
through the management port) even though ovn-k is not running.

This new health state allows interpreting this new scenario for which
other controllers might not want to unschedule their services. For
example, ovn-kube rollouts.

This new UNAVAILABLE health state is not supported in hypershift given
the implementation of proxied connections that we have there.

Signed-off-by: Jaime Caamaño Ruiz <jcaamano@redhat.com>