ipam, pod allocator: use PersistentAllocations interface
This way, the pod allocator has an uninitialized interface (which is nil)
instead of a nil implementation (which is not nil) as its ipam claims
reconciler attribute.

Previously, the claimsReconciler check in [0] would never be nil, which
would cause the controller to go look for the corresponding IPAMClaim
even if the feature was disabled, leading to a nil pointer - given the
claimsReconciler was nil.

[0] - https://github.com/ovn-org/ovn-kubernetes/blob/a31f23ffbab4fea4b4b1e2828ab53612db192bdb/go-controller/pkg/allocator/pod/pod_annotation.go#L293

Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>