Support OVS-DOCA
Offer the possibility to use ovs-doca based userland datapath in OVS
with OVN-kubernetes.

ovs-doca uses DOCA Flow library which enables use of the embedded
eswitch and other hardware based offload mechanisms in capable hardware.

The ovn bridge will be created with the 'netdev' datapath type.
ovnkube will query the datapath type of this bridge and modify OVS requests
as necessary. In particular, CNI requests resulting in ovs configuration will
add ports of type dpdk if on a 'netdev' bridge. Same with management
port as well.

For using ovs-doca with ovn-kubernetes, in addition to initializing
ovs-doca correctly, the following ovs configuration needs to be set

`ovs-vsctl set open . external-ids:ovn-bridge-datapath-type=netdev`

Signed-off-by: Hareesh Puthalath <hareeshp@nvidia.com>