Add QoSTable DBIndexing
This commit introduces DBIndexing for
QoSTable. Note that I have not added a
qos_syncer or anything to cleanup
older EQoS rules because this feature is
not GA yet, I don't see a need to support
upgrades and anyways we are planning to make
major changes to this feature moving forward.

Signed-off-by: Surya Seetharaman <suryaseetharaman.9@gmail.com>