Fixes waiting for level-driven informers to stop
We were not waiting on for informers to fully shutdown for level-driven
controllers. As a consequence, goroutines would still be running
handling events while the next unit test case ran. In this case, a
handler with no workqueue (like node_tracker) would attempt to read
global config, while the next test case was modifying the global config,
leading to a data race.

Fixes: #4381

Signed-off-by: Tim Rozet <trozet@redhat.com>