ANP: Allow two ANPs at the same priority
This commit allows creation of two ANPs
at the same priority. We will generate
an event for the user if two anps are
created at same priority. -> Only 1 event
per creation and during updating *if the
priority change leads to conflict; then we
do an event* -> Note to reviewers to check
the logic around event generation because
that is new.

It also breaks down a bigger UT into
3 smaller UTs to incorporate
previous review feedback for easiness
for reviewers

Signed-off-by: Surya Seetharaman <suryaseetharaman.9@gmail.com>