rm e2e-periodic and run e2e in full twice daily
this is to help track down and find the most common flakes we need
to fix. It removes the condition on the e2e job to not run on
a schedule so the cron will pick it up (twice daily). the e2e
job should still run pull_requests for [ master ] branch.

this also adds the "workflow_dispatch" trigger so we can manually
trigger the workflow from the github UI.

A small blurb about the e2e matrix running twice daily was added
to the ci.md doc as well as adding a blank line above each section
with bullets to make the .io site show them in bullet form instead
of all inline with each other.

Signed-off-by: Jamo Luhrsen <jluhrsen@gmail.com>