factory,control-manager: Add namespaces informer
The namespace active-network annotation is handled at the NAD
controller, since namespaces are common to the cluster they need to be
updated at a common places, in the case of IC this is control-manager.
This change add the namespace informer to it so nad controller can check
if the annotation is already set.

Signed-off-by: Enrique Llorente <ellorent@redhat.com>