Add primary field to pod annotation
This commit adds the primary network boolean
field to a pod's annotation info so that this
information can be used to tell if the pod
is managed by the default network OR
secondary network. This will be useful
in future commits and also for other feature
controller's to leverage in the future.

k8s.ovn.org/pod-networks: '{
  "default":{
    "ip_addresses":["10.244.1.4/24"],
    "mac_address":"0a:58:0a:f4:01:04",
    "routes":[{"dest":"10.244.0.0/16","nextHop":"10.244.1.1"},{"dest":"100.64.0.0/16","nextHop":"10.244.1.1"}],
    "ip_address":"10.244.1.4/24",
    "primary":false
  },
  "ns2/l2-network":{
    "ip_addresses":["10.100.200.9/24"],
    "mac_address":"0a:58:0a:64:c8:09",
    "ip_address":"10.100.200.9/24",
    "tunnel_id":2,
    "primary":true
  }
}'

Signed-off-by: Surya Seetharaman <suryaseetharaman.9@gmail.com>