Remove unused event recorder from NAD controller
The event recorder was used to inform that NAD updates were not allowed.
However the refactored NAD controller is able to handle NAD updates and
independently track the underlying network updates. Just as deleting and
readding a NAD, updating a NAD behavior is undefined if pods are
attached to the network and a webhook somewhere should be preventing
that.

Signed-off-by: Jaime Caamaño Ruiz <jcaamano@redhat.com>