Consider OVN networks with different topology than expected stale
On startup, networks with unrecognized names were considered stale and
cleaned up. These network were probably deleted while ovnkube was down.
But network configuration could also change during this time and stale
entities from the older configuration would not be cleaned up. Just as
we match on expected network names, match also on their topology as an
early effort to mitigate the problem.

As always, the behavior of network configuration changes while pods are
attached to it is undefined and a webhook somewhere should be preventing
that.

Signed-off-by: Jaime Caamaño Ruiz <jcaamano@redhat.com>