util: Return NetConf as active network
The user defined primary network feature need the NAD name of the
network to compose the NetworkSelectionElement, and the CNI part needs
the NetConf. This change refactor the
util.GetActiveNetwork into util.FindActiveNetwork that returns a pointer
to NetConf struct, this we we have nadName, networkName and NetConf, it also remove the
UnknownNetworkType since nil value states the same.

Signed-off-by: Enrique Llorente <ellorent@redhat.com>