[WIP FOR TESTING ONLY] Dockerfile.fedora: Use fedora rawhide image.
Fedora rawhide has OVN version - 23.03.90-1 which is the present OVN
main (with the commit [1]).  This is required to test new feature - xxx.

We can switch back to fedora 40 or the latest release once OVN 24.09 is
released and available in stable fedora image.

[1] - b0296698332a ("Maintainers: Fix the typo in the email id.")

Also removed installing python3-openvswitch and openvswitch-test packages
as it's failing to install these packages in rawhide with the below
error message

-----
Failed to resolve the transaction:
Problem: conflicting requests
  - nothing provides python3.12dist(sortedcontainers) needed by python3-openvswitch-3.3.0-1.fc41.x86_64 from rawhide
  - nothing provides python(abi) = 3.12 needed by python3-openvswitch-3.3.0-1.fc41.x86_64 from rawhide
You can try to add to command line:
  --skip-broken to skip uninstallable packages
-----

Signed-off-by: Numan Siddique <numans@ovn.org>