multi-homing, tests: secondary L2 nets do not auto-reserve IPs
This commit adds a unit test that ensures the IP allocator for secondary layer2
networks do not auto-reserve IPs for the gateway or logical network management
ports.

It also refactors a bit the organization of the current clustermanager test
structure, to make it clearer - i.e. sort out ginkgo contexts / when / it
blocks.

Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>