Remove BasicNetInfo, add MutableNetInfo and ReconcilableNetInfo
Consolidate all getters in NetInfo becoming the single interface to
obtain network information intended to be used broadly as read-only.

Consolidate all setters in new MutableNetInfo interface. A
MutableNetInfo can only be built from a copy of NetInfo. Add setters for
route advertisements related information. This is intended to be used by
a network manager to aggregate consolidated information about the
network from multiple sources that can change over time.

Add a ReconcilableNetInfo interface acting as a receiver by which
changes done in a MutableNetInfo can be incorporated into an existing
NetInfo. A ReconcilableNetInfo can only be built from a copy of NetInfo.
It is  intended to be used by top level network controllers to reconcile
changes to network information when instructed by a network manager.

New free methods IsNetworkCompatible, DoesNetworkNeedReconciliation, and
Reconcile facilitate the coordination of the reconciliation.

The overall idea is to have a tighter control of NetInfo changes and
making sure that controllers become aware of those changes in a way that
allows the controllers to reconcile them.

Signed-off-by: Jaime Caamaño Ruiz <jcaamano@redhat.com>