Zone & node network controllers reconcile NetInfo changes
Add the ability for network controllers to reconcile some network
information changes. Currently just changes of the VRFs the network is
leaking/advertising to. Support for reconciling NAD changes is not
included in this commit.

Currently reconciles:
- for zone network controllers to configure or not the pod IP to node IP
  SNAT on the GR for a nodes local to the zone
- for node network controllers to configure or not br-ex flows to
  redirect pod IP ingress traffic to the OVN network (except ingress to
  the management ip address). Only done for the default network, UDNs
  will be handled on a later commit.

This should be enough to provide direct ingress capabilities for the
default network in SGW mode.

Note that non-primary network controllers don't reconcile anything as
route advertising is not supported on them. Also cluster manager network
controllers don't reconcile much as they don't have the need.

Signed-off-by: Jaime Caamaño Ruiz <jcaamano@redhat.com>