L2/L3: Create UDN management-port on host
This commit creates the udn management port
on host side along with the OVS interface on
br-int.

NOTE: For L2 we need a central managementport
allocator that will be done in a separate PR.
Currently we use .2 on all nodes for a L2
network.

NOTE2: Add/Del Networks will pave way for more
additional things to be done in the future.

Signed-off-by: Surya Seetharaman <suryaseetharaman.9@gmail.com>