HELM: udn, non-IC: allow node to read NADs on non-IC
This is needed when there are user defined primary networks configured
in a namespace, and so the CNI side knows it needs to wait for the
annotations to appear on the pod.

This is a complement for commit 9e2aa33cf0bce0994ae96a1be982528ee16471a2
so it also applies to helm.

Fixes: #4534

Signed-off-by: Flavio Fernandes <ffernandes@nvidia.com>