udn, egress: introduce gateway topology manager
This commit refactors the existing code, moving the join switch /
cluster router creation away from the default network controller. This
is done because the layer2 controllers have no need for this type of
topology, and were able to to do previously. This will also make it
simpler to unit test the join switch / cluster router creation.

We ensure these logical entities have their respective network name in
the external IDs, so we a network controller can filter entities for the
network it manages.

The provided struct is properly unit tested.

Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>