Ensure config is reset before node ip handler tests
There was a panic seen in https://github.com/ovn-org/ovn-kubernetes/actions/runs/10077834705/job/27861284436?pr=4560

Due to the watchFactory starting and trying to list NADs would end up
with a NPE, as the client in node for unit test not having NAD support.
The suspect is that global config is not reset during these tests, and
UDN was enabled. This patch resets global config before the test runs to
ensure UDN is disabled as these tests do not require it.

Signed-off-by: Tim Rozet <trozet@redhat.com>