Rule manager: delete rule even if not managed
While testing IP rule deletion, rule
manager will not attempt to delete an IP rule if it doesn't
manage it. This makes sense for efficiency but after a restart
when the managed IP rule cache is reset, callers may cleanup rules
and wish for rules to be deleted.

Thankfully, the consumer (EIP) of rule manager called "OwnPriority" for
a particular rule priority, therefore stale rules were cleaned up anyway because
any unknown rules are deleted if they arent managed, therefore theres
no bug in the product.

Signed-off-by: Martin Kennelly <mkennell@redhat.com>