udn, l2: Use conditional SNAT
At primary UDN for layer2 the gateway router is connected directly to
the network switch, so traffic comming from the gateway router cannot be
SNATed with masquerade IP since pods will not able to answer it back.

This change use conditional SNAT to masquerade only the traffic that
goes out of the cluster using the rtoe logical port.

Signed-off-by: Enrique Llorente <ellorent@redhat.com>