util, cni: Extract validation logic from parsing NAD code
Enable reuse the NAD spec validation logic that is part of the
prase NAD code, in a way a structured NAD spec can be validated
instead of raw JSON string.

The NAD templating code will be introduced in a follow up commit
and make use of the common NAD validation code.

This is a preliminary step before introducing the NAD templating code,
enable it to validate given NAD spec before creating the actual NAD.

Signed-off-by: Or Mergi <ormergi@redhat.com>