default net bootstrap: ignore localnet patch ports
This commit exports the code that find the bridge to which the default
cluster network patch port is attached into a function.

This way, we can properly unit test it, ensuring it works against the
patch ports we know will be present in the system - secondary localnet
networks, and primary UDN patch ports.

Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>