udn/nad e2e testing ip family aware
currently the udn/nad primary network e2e testing does nothing to check
to state of the cluster before creating the network. This makes it
possible to test primary networks with ip families that the underlying
cluster does not support which is not possible.

This commit ensures that e2e testing will only create primary networks
that conform to the cluster being tested

Signed-off-by: Jacob Tanenbaum <jtanenba@redhat.com>