Fixes deleting CNI DEL with UDN
The previous implementation had several issues:
 - If a NAD was removed while pods were active, CNI DEL would fail.
 - If a pod was force removed, CNI DEL would fail.

The above issues were due to:
 - Searching for the NAD to be able to do the CNI DEL
 - Searching for the pod in the informer cache for CNI DEL

This PR simplifies deletion by not relying on the pod object or the NAD
object. When primary CNI deletion happens with this patch, we simply
find any interfaces that belong to this pod's sandbox ID. There is no
performance impact, because clearPodBandwidth was already finding all of
these ports anyway.

CNI deletion happens in reverse order, so we can be sure that secondary
interfaces have already been deleted by the time we search and delete
interfaces for the primary CNI.

Signed-off-by: Tim Rozet <trozet@redhat.com>