Make GetActiveNetworkForNamespace use a controller
Leverages the NAD Controller to also track primary networks. This should
be faster than listing all NADs in a namespace and iterating through
them to find the primary network for a namespace.

Refactors network manager and nad controller. Changes the network cache to
only be stored in network manager, rather than having a duplicate cache in
NAD controller.

NAD Controller is modified to only hold references to network names for
NADs, and then queries network manager to get the networks. NAD Controller
and Network Manager are now thread safe.

Tests are also modified:
 - NAD Controller now implemented as an interface so that unit tests can
   implement a fake NAD controller for getting active network
 - Fully functioning OVN controller unit tests still leverage a fully
   functioning NAD controller. NCM may be faked out in these cases.
 - Fake entities moved to testing/nad package

Co-authored-by: Jaime Caamaño Ruiz <jcaamano@redhat.com>
Signed-off-by: Tim Rozet <trozet@redhat.com>